+++ CHANGELOG.md
@@ -16,6 +16,7 @@ omnibus-gitlab repository.
 - Update docker registry to include a set of patches from the upcoming 2.7.0 release !2888
 - Update Mattermost to 5.5.0
 - Add impersonation_enabled configuration to gitlab.rb !2880
+- Update runit version to 2.1.2 !2897
 
 11.5.2